requirements.txt
Flask
pymongo
pdfplumber
python-dotenv
groq
flask-bcrypt
gunicorn
fpdf
pdfkit